complete problem management

+++ FU.OJ.Server/Controllers/ProblemController.cs
@@ -1,7 +1,6 @@
 ﻿using FU.OJ.Server.DTOs;
 using FU.OJ.Server.DTOs.Problem.Request;
 using FU.OJ.Server.Infra.Const.Route;
-using FU.OJ.Server.Infra.Models;
 using FU.OJ.Server.Service;
 using Microsoft.AspNetCore.Mvc;
 
@@ -13,10 +12,14 @@ namespace FU.OJ.Server.Controllers
     public class ProblemController : BaseController
     {
         private readonly IProblemService _service;
+        private readonly ITestcaseService _testcaseService;
 
-        public ProblemController(IProblemService service, ILogger<ProblemController> logger) : base(logger)
+
+        public ProblemController(IProblemService service, ILogger<ProblemController> logger, ITestcaseService testcaseService) : base(logger)
         {
             _service = service;
+            _testcaseService = testcaseService;
+
         }
 
         [HttpPost(ProblemRoute.Action.Create)]
@@ -66,11 +69,11 @@ namespace FU.OJ.Server.Controllers
         }
 
         [HttpPut("update")]
-        public async Task<IActionResult> UpdateProblemAsync(string id, [FromBody] UpdateProblemRequest request)
+        public async Task<IActionResult> UpdateProblemAsync([FromBody] UpdateProblemRequest request)
         {
             try
             {
-                var updated = await _service.UpdateAsync(id, request);
+                var updated = await _service.UpdateAsync(request);
 
                 if (!updated)
                     return NotFound();
@@ -88,10 +91,10 @@ namespace FU.OJ.Server.Controllers
         {
             try
             {
-                var deleted = await _service.DeleteAsync(id);
-
-                if (!deleted)
-                    return NotFound();
+                var problem = await _service.GetByIdAsync(id);
+                if (problem == null) return NotFound();
+                await _testcaseService.DeleteAsync(id);
+                await _service.DeleteAsync(id);
 
                 return NoContent();
             }
